Provide bamstat workaround option

+++ scripts/module3.sh
@@ -11,26 +11,40 @@ params=$2
 #Source params file
 . ${params}
 
-#Submit bamstat job per sample
-while read ID bam sex; do
-  bsub -q normal -o ${OUTDIR}/logs/bamstat.log -e ${OUTDIR}/logs/bamstat.log -sla miket_sc -J ${COHORT_ID}_bamstat "${liWGS_SV}/scripts/bamstat.sh -n -s 3 -i ${WRKDIR}/${ID}/${ID}.bam -o ${WRKDIR}/${ID}/bamstat/" 
-done < ${samples_list}
+#Skips running bamstat if ${pre_bamstat}==TRUE
+if ! [ ${pre_bamstat}=="TRUE" ] || ! [ -e ${bamstat_paths} ]; then
 
-#Gate until complete; 20 sec check; 5 min report
-GATEcount=$( bjobs -w | awk '{ print $7 }' | grep -e "${COHORT_ID}_bamstat\|BAMSTAT_GATE" | wc -l )
-GATEwait=0
-until [[ $GATEcount == 0 ]]; do
-  sleep 20s
+  #Submit bamstat job per sample
+  while read ID bam sex; do
+    bsub -q normal -o ${OUTDIR}/logs/bamstat.log -e ${OUTDIR}/logs/bamstat.log -sla miket_sc -J ${COHORT_ID}_bamstat "${liWGS_SV}/scripts/bamstat.sh -n -s 3 -i ${WRKDIR}/${ID}/${ID}.bam -o ${WRKDIR}/${ID}/bamstat/" 
+  done < ${samples_list}
+
+  #Gate until complete; 20 sec check; 5 min report
   GATEcount=$( bjobs -w | awk '{ print $7 }' | grep -e "${COHORT_ID}_bamstat\|BAMSTAT_GATE" | wc -l )
-  GATEwait=$[${GATEwait} +1]
-  if [[ $GATEwait == 15 ]]; then
-    echo "$(date): INCOMPLETE"
-    echo "$(date): Waiting on ${GATEcount} jobs to complete"
-    GATEwait=0
-  fi
-done
+  GATEwait=0
+  until [[ $GATEcount == 0 ]]; do
+    sleep 20s
+    GATEcount=$( bjobs -w | awk '{ print $7 }' | grep -e "${COHORT_ID}_bamstat\|BAMSTAT_GATE" | wc -l )
+    GATEwait=$[${GATEwait} +1]
+    if [[ $GATEwait == 15 ]]; then
+      echo "$(date): INCOMPLETE"
+      echo "$(date): Waiting on ${GATEcount} jobs to complete"
+      GATEwait=0
+    fi
+  done
 
-#Remove unnecessary pairs files
-while read ID bam sex; do
-  rm ${WRKDIR}/${ID}/bamstat/*pairs*txt
-done < ${samples_list}
+  #Remove unnecessary pairs files
+  while read ID bam sex; do
+    rm ${WRKDIR}/${ID}/bamstat/*pairs*txt
+  done < ${samples_list}
+
+else
+
+  #copy clusters and stats file from previous bamstat run
+  while read ID bam sex; do
+    mkdir ${WRKDIR}/${ID}/bamstat/
+    bpath=$( fgrep -w ${ID} ${bamstat_paths} | cut -f2 )
+    cp ${bpath}/*clusters* ${bpath}/stats.file ${WRKDIR}/${ID}/bamstat/
+  done < ${samples_list}
+
+fi
